Change Homestuck BETA to Homestuck, update URL
Homestuck moved to homestuck.com in April 2018. Also, while the non-beta comic is not entirely made in Flash, Flash is still a core part of it. You'd be missing a very large part of the experience if you skipped all content made in Flash, and the comic wouldn't even be readable on the old site due to some links being inside what were formerly Flash movies.